Moved WebcastListItem to separate file; added props validation
